Reference and Component identity equality refactoring
Refactor the DataStructure and Component so that the operations can
rely on reference equality instead of names. This allows to write more
elegant implementations of the Keep, Drop, Filter, Fold and Unfold operations.

The DataStructure is now striclty immutable and operations need to re-use previous
instances so that the scope can keep track of the changes. For instance, for the drop
operation:

DataStructure.Builder newDataStructure = DataStructure.builder();
for (Map.Entry<String, Component> componentEntry : dataset.getDataStructure().entrySet()) {
    Component component = componentEntry.getValue();
    if (!components.contains(component) || component.isIdentifier()) {
        newDataStructure.put(componentEntry);
    }
}
return newDataStructure.build();