Changes the JoinSpliterator so that the merger decide wether or not a tuple is generated
In order to implement the outer joins, the merger needs to be able to decide if a tuple
should be passed for each pairs. This is now implemented in the form of a TriFunction that
takes the left and right tuples, the result of the comparison and return a nullable tuple.

Another approach could have been to pass the Consumer of the Spliterator, but it seemed a bit
more complex to implement at this point.