Fix unnecessary call to getData() when ordering is available
The AbstractJoinOperation uses the Optional<Stream<Datapoint>> getData(Order order) method
to optimize the merge operation. If the returned optional is not empty there should not be
any other call to getData() since it could lead to non closed streams.

This is solved by using the orElseGet(Supplier<T> supplier) instead of the orElse(<T> value).