Manually close the undelying stream when closing StreamUtils.aggregate() streams
The protonpack library we are using to aggregate does not close the streams it is
using when the returned stream is closed.
I created a pull request https://github.com/poetix/protonpack/pull/42 to fix this
in the library. Until it is merged, we need to manually close the stream used to
create aggregations.